fix: use floating-point accumulator in sum_functions for integer grids
np.zeros_like(x) inherits the dtype of the union grid, so when the grid is
integer-valued the accumulator cannot hold the floating-point results of
combined functions such as Polynomial, raising UFuncTypeError. Initialize
the accumulator with a float dtype, restored from the prior behavior of
sum(f(x) for f in funcs) which promoted to float.

Adds a regression test combining an integer-valued tabulated function
with a polynomial (test_sum_functions_integer_grid).

+++ openmc/data/function.py
@@ -48,8 +48,11 @@ def sum_functions(funcs):
 
         # Evaluate each function and add together.  Tabulated functions are
         # only evaluated where they are defined; values beyond a function's
-        # tabulated range do not contribute to the sum.
-        y = np.zeros_like(x)
+        # tabulated range do not contribute to the sum.  The accumulator uses
+        # a floating-point dtype since the combined functions (e.g.,
+        # polynomials) may return values that cannot be represented
+        # losslessly in the dtype of an integer-valued grid.
+        y = np.zeros_like(x, dtype=float)
         for f in funcs:
             if isinstance(f, Tabulated1D):
                 within = ((x >= f.x[0]) & (x <= f.x[-1])) | \
